pyproject.toml
[tool.poetry]
name = "website-data-finder"
version = "0.1.0"
description = ""
authors = ["zaidejaz <zaidejaz521@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.2"
colorlog = "^6.8.2"
gspread = "^6.1.0"
oauth2client = "^4.1.3"
lxml = "^5.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
